Task 1 Submission
In this implementation of DNAHasher (task 1), we are given seq as a parameter
this allows us to calculate the length of k we use size().
To get prefix and suffix, we calculate the length of the prefix and suffix by
diving k by 2. If kmer is even then this obviously works, but if kmer is odd
then since we're using an int we will round down and we do k - prefix_len which
should give us the suffix length. Then we combine using bit shift and OR and
return that value as the hash key.

Signed-off-by: Raghda Kailany <raghdakailany@Raghdas-MacBook-Pro.local>